Create Event Source
An EventSource is an Akka Streams Source that reads events from an
EventReader and emits those events downstream on demand.
The source may be materialized to a SourceQueue which can be used
to signal source updates. If downstream demand exists and no source
updates are emitted through the SourceQueue, periodic ticks are used
to query for event updates.

A GapDetection mechanism is implemented which uses an event timestamp
to detect and persist gaps.

Resolves: LOG-25