Create SequencedMessage envelope for arbitrary message types
The generic `SequencedMessage` class is used to enrich arbitrary payloads with
metadata regarding the payload source and sequence-number.
All payloads that are serializable by an Akka `StringManifestSerializer` can be
wrapped within a `SequencedMessage`. Other serializer types are not supported.

An Akka serializer which uses protobuf is supplied for `SequencedMessage`.

A Kafka serializer which uses the aformentioned Akka serializer is
provided for `SequenceMessage`. This serializer can only be used from code
as an `ActorSystem` has to be provided at instantiation time and may therefore not
be used as a Kafka configuration property.