Create Transactional Event Producer
A `TransactionalEventProducer` is a Kafka message producer which reads
arbitrary events from supplied a transactional `EventStore` and creates
instances of `SequencedEvent` containing the event.

Events can be  persisted through an `EventWriter` which is supplied on
start of the `TransactionalEventProducer`.

Committed events are serialized by the producer and written to a Kafka topic
where the target-topic can be configured per event.
For each event an aggregate-id is extracted which is used as the key of the
`ProducerRecord` written to Kafka in order to guarantee the ordering of messages
per aggregate.