Producer should block indefinitely on exhausted Write-Buffer
According to various sources and local tests, messages may be dropped
if the write-buffer of a producer is exhausted and the duration specified
in the configuration setting `max.block.ms` has elapsed as the given
message will be considered failed.
Subsequent messages may be sent successfully if the write-buffer
was drained before the producer-stream was shutdown introducing message
re-ordering or lost messages if the previously failed messages are not retried.

To mitigate this issue the configuration setting `max.block.ms` is set to
Long.MAX_VALUE which blocks the producer indefinitely on a full write-buffer.