Add DelegatingStringManifestPayloadSerializerWithTombstoneDeserializer
When an event is not deserializable then consumers should be able to able to have a choice. The default behaviour for the PayloadFormatSerialzer is to throw an exception. Add a DelegatingStringManifestPayloadSerializerWithTombstoneDeserializer to the PayloadFormatSerializer which will deserialize failed events into a Tombstone class object. This Tombstone contains the original serializer id and payload string manifest as well as the bytes of the original payload.