Consumer should be able to decide what to do with payloads which are not deserializable.
By default, if an exception is thrown when deserialising a payload this will not be matched and the Consumer Stream will be stopped.

Adding a additional factory methods for the PayloadFormatDeserializer will enable the user to somewhat react to such payloads. The PayloadFormatDeserializer insead of returning an [A] returns a [Try[A]] instead. This way for example the user could log out the cause of the failing deserialization and continue consuming the topic or the user could only filter certain causes if need may be.

A use case could be that an updated producer is emitting new events whereas the consumer will be updated later or potentially never.