Deletion of aggregated Event-Batches is not working as intended
The `TransactionalEventProducer` should delete events in batches defined by a configured time-interval.
Delete-requests are performed with the sequence-number of the last event within a batch.
The current implementation however does not batch delete-requests but issues a delete-request per event which
puts unncessary pressure on the underlying `EventStore`. This behaviour only occurs using Akka versions prior to 2.5.

To mitigate this issue in Akka 2.4 the `delay` modifier is used to delay a delete-request for an event
and merge all delete-requests occuring within the interval of a single delay, preserving only the latest event.