Add Lifecycle-methods to TransactionalEventProducer (#15)
In the current version of the `TransactionalEventProducer` creation of an `EventWriter` is necessarily preceded
by the start of the the event-production. This behaviour may not be desired as a client may only want to store
events to the underlying `EventStore` without actual event-production to Kafka.
Addtionally a `TransactionalEventProducer` may only be started and stopped exactly once in the current implementation,
complicating situations where a client wants to perform these operations mulitple times
(e.g. in case of a single event-producer with leader-election).

To accomodate these situations the following changes are introduced:
- Extend the `TransactionalEventProducer` by additional methods allowing clients to
	* create a new instance of an `EventWriter`
	* `stop()` the `TransactionalEventProducer`, stopping all event production
	* `terminate()` the `TransactionalEventProducer`,  stopping event production and cleaning all used resources
- Allow the `TransactionalEventProducer` to be restarted after the `stop()` method has been called.

Closes: LOG-93