Create Event-Consumer (#18)
Event-Consumer:
Create `EventConsumer` to consume events from Kafka and invoking an `EventHandler` for each consumed event.
After successful handling of an event the offset is committed to Kafka.
Uses the `StreamExecutor` to run the underlying graph and only stop the graph upon encountering an event which can
not be deserialized.

Deduplication-Batch:
Implement a `DeduplicationBatch` which deduplicates entries according to their aggregate-ids per topic and partition.
Drops all entries with the same aggregate-id entries already contained in the batch per topic and partition.
Preserves the order of entries per a partition.

StreamExecutor:
Implement a `StreamExecutor` which is supplied with an Akka streams `Graph`, materializes
the stream and keeps the stream executed in case of errors based on a configurable
supervision-strategy.
Supervision supports an expoential backoff algorithm.